Split too large stampit-specs.js to the several files.
